+++ example/CHANGELOG.md
@@ -39,5 +39,5 @@
 #### 0.3.3
 * 新增 Menu 小红点功能
 * 新增 清理缓存 功能
-
+* 删除 getChildFragment().getFragments() 方法